Move extended_enum to utilities package
