+++ sous-chef/sous_chef/abstract/handle_exception.py
@@ -2,7 +2,8 @@ from dataclasses import dataclass, field
 from typing import Callable, List, Tuple
 
 from omegaconf import DictConfig
-from sous_chef.abstract.extended_enum import ExtendedEnum
+
+from utilities.extended_enum import ExtendedEnum
 
 
 @dataclass
